Restyle the sign in/sign up modal to show sign in by default Fixes #109
