+++ src/components/Modals/Sign_in.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useRef } from "react";
 import { Modal, Button } from "react-bootstrap";
 import "./Modals.css";
 import "bootstrap/dist/css/bootstrap.min.css";
@@ -37,17 +37,12 @@ function Sign_in(props) {
     // Sign into Firebase using popup auth & Google as the identity provider.
     auth.signInWithPopup(facebookk);
   };
-  const handleClose = () => {
-    setShow(false);
-    props.showSignIn();
-  };
-
-  useEffect(() => {
-    setShow(props.show);
-  }, [props.show]);
+  const handleClose = () => setShow(false);
+  const handleShow = () => setShow(true);
 
   return (
     <>
+      <div onClick={handleShow}>SignIn</div>
       {/* Modal header */}
 
       <Modal show={show} onHide={handleClose}>
@@ -85,7 +80,7 @@ function Sign_in(props) {
               </div>
               <div className="col-10 col-sm-10 col-md-10 p-0 my-auto mx-auto ">
                 <input
-                  type="text"
+                  type="password"
                   name="report"
                   placeholder="Password"
                   className="ModalBody"
@@ -101,9 +96,10 @@ function Sign_in(props) {
               <div className="col-12 col-sm-12 col-md-12 text-center">
                 <Button
                   id="SignInButton"
-                  onClick={() =>
-                    sign_In(emailRef.current.value, passwordRef.current.value)
-                  }
+                  onClick={() => {
+                    sign_In(emailRef.current.value, passwordRef.current.value);
+                    handleClose();
+                  }}
                 >
                   Sign In
                 </Button>
@@ -133,6 +129,20 @@ function Sign_in(props) {
                 </a>
               </div>
             </div>
+          </div>
+          {/* SIGN Up */}
+
+          <div className="row text-center">
+            <p>Do not have an account?</p>
+            <a
+              href="#"
+              onClick={() => {
+                props.showSignUp();
+                handleClose();
+              }}
+            >
+              <p>Sign Up</p>
+            </a>
           </div>
         </Modal.Body>
       </Modal>
